requirements.txt
# Scrapping
beautifulsoup4==4.6.3
pyppeteer==1.0.2
nest_asyncio==1.5.4
asyncio==3.4.3
pandas==1.4.0
requests==2.27.1

# DataBase
SQLAlchemy==1.4.31
psycopg2-binary==2.9.3

# Mock data
Faker==11.3.0

# Utils
greenlet==1.1.2
pipe==1.6.0
python-dateutil==2.8.2
six==1.16.0
text-unidecode==1.3
pyyaml==6.0

